requirements.txt
annotated-types==0.7.0
blinker==1.8.2
blis==0.7.11
Brotli==1.1.0
camelot-py==0.9.0
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.5
contourpy==1.2.1
cryptography==42.0.7
cssselect2==0.7.0
cycler==0.12.1
cymem==2.0.8
distro==1.9.0
docopt==0.6.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
et-xmlfile==1.1.0
Faker==25.2.0
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.52.1
greenlet==3.0.3
html5lib==1.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
JPype1==1.5.0
kiwisolver==1.4.5
langcodes==3.4.0
langdetect==1.0.9
language_data==1.2.0
marisa-trie==1.1.1
MarkupSafe==2.1.5
matplotlib==3.9.0
murmurhash==1.0.10
mysql-connector-python==8.4.0
nltk==3.8.1
num2words==0.5.13
numpy==1.26.4
opencv-python==4.9.0.80
openpyxl==3.1.2
openpyxl-image-loader==1.0.5
packaging==24.0
pandas==2.2.2
pdf2image==1.17.0
pdfkit==1.0.0
pdfminer.six==20231228
pdfplumber==0.11.0
pillow==10.3.0
preshed==3.0.9
pycparser==2.22
pydantic==2.7.2
pydantic_core==2.18.3
pydyf==0.10.0
PyMuPDF==1.24.4
PyMuPDFb==1.24.3
pyparsing==3.1.2
PyPDF2==3.0.1
pypdfium2==4.30.0
pyphen==0.15.0
pytesseract==0.3.10
python-dateutil==2.9.0.post0
pytz==2024.1
regex==2024.5.15
reportlab==4.2.0
requests==2.32.3
setuptools==70.0.0
six==1.16.0
smart-open==6.4.0
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.30
srsly==2.4.8
tabula-py==2.9.3
thinc==8.2.3
tinycss2==1.3.0
tk==0.1.0
tqdm==4.66.4
typer==0.9.4
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
wasabi==1.1.3
weasel==0.3.4
weasyprint==62.1
webencodings==0.5.1
Werkzeug==3.0.3
wtf==0.1
WTForms==3.1.2
XlsxWriter==3.2.0
zopfli==0.2.3
